Move single subscription delete button into SubscriberInfo directory
